Project Alpha/requirements.txt
# Core Trading
pandas>=1.3.0
TA-Lib>=0.4.0
requests>=2.28.0
python-dotenv>=0.20.0

# Brokers
deriv-api>=1.0.0
MetaTrader5>=5.0.0

# Web Dashboard (Dash)
plotly>=5.0.0
dash>=2.0.0
dash-bootstrap-components>=1.0.0

# Mobile App (Kivy)
kivy>=2.2.0
kivy-garden>=0.1.5
matplotlib>=3.7.0

# Backend API
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.0.0
python-multipart>=0.0.6

# Deployment
docker>=6.0.0
docker-compose>=1.29.0

# Utilities
numpy>=1.24.0
aiofiles>=23.0.0
websockets>=11.0.0
